Update and rename f33.py to f34.py

+++ examples/text_to_image/f34.py
@@ -1058,7 +1058,7 @@ def main():
     param_spec = jax.tree_util.tree_map(lambda x: partition_shape(x.shape) , params )
 
     text_params = jax.tree_util.tree_map(lambda x: jax.device_put(x ,NamedSharding(mesh , partition_shape(x.shape)) ).astype(jnp.bfloat16), text_params)
-    unet_params = jax.tree_util.tree_map(lambda x: jax.device_put(x ,NamedSharding(mesh , partition_shape(x.shape)) ).astype(jnp.float32), params)
+    unet_params = jax.tree_util.tree_map(lambda x: jax.device_put(x ,NamedSharding(mesh , partition_shape(x.shape)) ).astype(jnp.bfloat16), params)
     
 #     opt_state = optimizer2.init(unet_params)
 #     opt_state = jax.tree_util.tree_map(lambda x: jax.device_put(x ,NamedSharding(mesh , partition_shape(x.shape)) ), opt_state)
